+++ src/errors/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Errors Module
+ *
+ * @packageDocumentation
+ */
+
+export { BluetoothPrintError, ErrorCode } from './BluetoothError';
+export { ConnectionError, ConnectionErrorCode } from './ConnectionError';
+export { PrintJobError, PrintJobErrorCode } from './PrintJobError';
+export { CommandBuildError, CommandBuildErrorCode } from './CommandBuildError';