+++ src/factory/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Factory Module
+ *
+ * @packageDocumentation
+ */
+
+export {
+  createBluetoothPrinter,
+  createWebBluetoothPrinter,
+  PrinterFactory,
+  type PrinterFactoryOptions,
+} from './PrinterFactory';